Break Artemis devices methods into seperate files
